refactored slighltly
